Make attendee limit big int

+++ database/migrations/2020_09_18_120549_add_attendee_limit_column_to_social_mobs_table.php
@@ -9,7 +9,7 @@ class AddAttendeeLimitColumnToSocialMobsTable extends Migration
     public function up()
     {
         Schema::table('social_mobs', function (Blueprint $table) {
-            $table->unsignedInteger('attendee_limit')->default(PHP_INT_MAX);
+            $table->unsignedBigInteger('attendee_limit')->default(PHP_INT_MAX);
         });
     }
